Support chrooted builds.
This adds a chroot handling script and a few additional variables to
pbulk.conf:

  - $chroot_create and $chroot_delete point to scripts which create and
    delete a sandbox respectively, each taking a path argument of the
    chroot location.

  - $chroot_dir is the base directory where chroots are created.

  - $build_chroots and $scan_procs specify the number of build chroots
    to create and scan processes to run.

If the number of build chroots is 1, it is named:

  ${chroot_dir}-build

otherwise they are named i=1..N:

  ${chroot_dir}-build-${i}

For scan processes a single ${chroot_dir}-scan chroot is created and the
requested number of processes are started inside it.

The chroot_dir variable isn't brilliantly named, but at this point is
baked into too many of our pbulk branches to change.

Currently users of this need to provide their own create/delete scripts,
though hopefully at some point I'll provide sysutils/mksandbox wrappers
as examples for getting started.