mk: Remove _GCC_REQD pmatch check.
This generally adds quite a bit of complexity, badly named variables,
and increases costs by having to run pkg_admin pmatch for every
operation for what is essentially just a check that GCC_REQD isn't
higher than what's available in pkgsrc (which it should never be).

This is much better suited to a pkglint check or a simpler pattern
match.