mk: Simplify _GCC_VERSION logic.
Remove support for egcs, it is beyond obsolete at this point and
wouldn't be able to bootstrap even our oldest GCC.

Remove DRAGONFLY_CCVER support.  Again these are very obsolete versions
of GCC, and anyone wanting to use an obsolete version of DragonFly would
use either of them to first bootstrap to a modern pkgsrc GCC, or can
simply set CCVER in their environment.

Remove check for 'gcc version'.  There's no merit in running additional
commands for every make invocation.  If this file has been included then
we are assuming the user wishes to use GCC, and -dumpversion will detect
whether it is suitable or not.

Remove _GCC_PKG variable and simply inline its contents.  It was badly
named as it is the builtin GCC version, and not in any way a package.