mk: Overhaul LANGUAGES.gcc setup.
Remove the pointless copy paste of supposed supported languages, when we
never actually test to see if that language is supported by that GCC.
It is entirely possible that the user has opted to build without support
for certain languages.

Instead just list all of the languages that modern GCC supports.  If for
whatever reason a user builds a stripped-down GCC then they can now set
LANGUAGES.gcc themselves, and for reasonable cases where newer languages
are required we can simply use GCC_REQD to indicate which version
introduced support for them.