mk: Add IMPLICIT_DEPENDS and stricter check-shlibs.
This helps to expose potential issues where buildlink does not have the
full graph of dependencies due to inclusion guards, and libraries may be
buildlinked but are only build dependencies, causing runtime failures
that are not caught by the regular check-shlibs.

The situation was muddied slightly by our builds of GCC defaulting to
search for includes and libraries under LOCALBASE, but that has been
resolved by enhancing the custom specs file used by gcc-libs to remove
those paths.

This isn't ready for upstream yet as it may cause false positives due to
buildlink's incomplete graph, but we want to be alerted to potential
problems so that we can investigate and fix, and we would rather see
false positives than potentially let broken packages out undetected.