ncurses: Update to 6.5 and overhaul package.
This combines ncurses and ncursesw into a single package, with the wide
character version becoming the default and only implementation.

Some applications that use very old configure scripts still assume that
only ncursesw contains support for wide characters, and so an extra
BUILDLINK_TRANSFORM is provided to redirect libncursesw requests to
libncurses.  This is enough to fix e.g. security/pinentry.

A number of other cleanups are included:

 - stop using a custom TERMINFODIR on SunOS
 - remove unused PLIST_SUBST
 - use newer make syntax

This package now conflicts with and supersedes ncursesw-[0-9]*.  Removal
of support for devel/ncursesw will come next.